+++ x11/x11.cu
@@ -133,9 +133,9 @@ extern "C" int scanhash_x11(int thr_id, uint32_t *pdata,
 	const uint32_t first_nonce = pdata[19];
 
 	int intensity = (device_sm[device_map[thr_id]] > 500) ? 256 * 256 * 19 : 256 * 256 * 10;
-	int throughput = opt_work_size ? opt_work_size : intensity; // 20=256*256*16;
+	uint32_t throughput = opt_work_size ? opt_work_size : intensity; // 20=256*256*16;
 
-	throughput = min(throughput, (int)(max_nonce - first_nonce));
+	throughput = min(throughput, max_nonce - first_nonce);
 
 
 	if (opt_benchmark)
